# sql: Make cancel checker more efficient