--- a/pkg/sql/colexec/colexechash/hash_utils.eg.go
+++ b/pkg/sql/colexec/colexechash/hash_utils.eg.go
@@ -1088,7 +1088,7 @@ func rehash(
 	default:
 		colexecerror.InternalError(errors.AssertionFailedf("unhandled type %s", col.Type()))
 	}
-	cancelChecker.CheckEveryCall()
+	cancelChecker.Check()
 }
 
 // rehash64 takes an element of a key (tuple representing a row of equality
@@ -2148,5 +2148,5 @@ func rehash64(
 	default:
 		colexecerror.InternalError(errors.AssertionFailedf("unhandled type %s", col.Type()))
 	}
-	cancelChecker.CheckEveryCall()
+	cancelChecker.Check()
 }