--- a/pkg/sql/colexec/colexechash/hash_utils_tmpl.go
+++ b/pkg/sql/colexec/colexechash/hash_utils_tmpl.go
@@ -159,7 +159,7 @@ func rehash(
 	default:
 		colexecerror.InternalError(errors.AssertionFailedf("unhandled type %s", col.Type()))
 	}
-	cancelChecker.CheckEveryCall()
+	cancelChecker.Check()
 }
 
 // rehash64 takes an element of a key (tuple representing a row of equality
@@ -206,5 +206,5 @@ func rehash64(
 	default:
 		colexecerror.InternalError(errors.AssertionFailedf("unhandled type %s", col.Type()))
 	}
-	cancelChecker.CheckEveryCall()
+	cancelChecker.Check()
 }