--- a/pkg/sql/rowexec/zigzagjoiner.go
+++ b/pkg/sql/rowexec/zigzagjoiner.go
@@ -387,7 +387,7 @@ func (z *zigzagJoiner) Start(ctx context.Context) {
 		ctx, zigzagJoinerProcName, &z.contentionEventsListener,
 		&z.scanStatsListener, &z.tenantConsumptionListener,
 	)
-	z.cancelChecker.Reset(ctx, rowinfra.RowExecCancelCheckInterval)
+	z.cancelChecker.Reset(ctx)
 	log.VEventf(ctx, 2, "starting zigzag joiner run")
 }