--- a/pkg/util/ctxutil/context_abi_pre1_20.go
+++ b/pkg/util/ctxutil/context_abi_pre1_20.go
@@ -1,0 +1,38 @@
+// Copyright 2023 The Cockroach Authors.
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.txt.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0, included in the file
+// licenses/APL.txt.
+
+//go:build !go1.20
+
+package ctxutil
+
+import (
+	"context"
+	_ "unsafe" // Must import unsafe to enable gross hack below.
+)
+
+// Gross hack to access internal context function.  Sometimes, go makes things
+// SO much more difficult than it needs to.
+
+// A canceler is a context type that can be canceled directly. The
+// implementations are *cancelCtx and *timerCtx.
+type canceler interface {
+	cancel(removeFromParent bool, err error)
+	Done() <-chan struct{}
+}
+
+//go:linkname context_propagateCancel context.propagateCancel
+func context_propagateCancel(parent context.Context, child canceler)
+
+//go:linkname context_removeChild context.removeChild
+func context_removeChild(parent context.Context, child canceler)
+
+func (c *whenDone) cancel(removeFromParent bool, err error) {
+	c.cancelWithCause(removeFromParent, err, nil)
+}