Arrange for cancel checker to be notified once context
is done.

```
Old implementation:
BenchmarkCancelChecker-10    	133185742	   8.790 ns/op
New Implementation:
BenchmarkCancelChecker-10    	1000000000	 0.5804 ns/op
```

Epic: None

Release note: None