@miretskiy 2023-05-31
> nit: I'd extract this comment from the code into the commit message (I don't think we usually put things like this into the code).

done