@shermanCRL 2023-05-31
I’d be interested @yuzefovich to know if this would be a visible improvement in SQL performance? Any sort of macro benchmark that may be interesting?