@miretskiy 2023-05-31
Yeah; the overall change is not great; colexec calls to CheckEvery might have been more expensive... but still, it's hard to see those if you're talking seconds.  A solid maybe :)