@shermanCRL 2023-05-31
If you don’t foresee visible macro improvements, doesn’t sound like a good use of your time @yuzefovich! Just curious. If it were in a hot loop the answer might be different. Thanks.