@miretskiy 2023-06-01
Probably due to uint32* allocation. And that's because planner copied
canceler. If it's guaranteed that reset is called, this alloc can be removed

On Wed, May 31, 2023, 4:59 PM Yahor Yuzefovich ***@***.***>
wrote:

> Interestingly, this seems to lead to an increase in allocations which
> actually increases the latency on the hot path:
>
> ***@***.***:~/go/src/github.com/cockroachdb/cockroach$ BENCHTIMEOUT=72h PKG=./pkg/sql/flowinfra BENCHES=BenchmarkFlowSetup ./scripts/bench HEAD~1 HEAD~0
> Comparing HEAD~0 (new) with HEAD~1 (old)
>
> Writing to /tmp/tmp.Y7b21cdmu3
> Switching to HEAD~1
> + ./dev bench ./pkg/sql/flowinfra --timeout=72h --filter=BenchmarkFlowSetup --count=30 --bench-mem -v --stream-output --ignore-cache
> Switching to HEAD~0
> + ./dev bench ./pkg/sql/flowinfra --timeout=72h --filter=BenchmarkFlowSetup --count=30 --bench-mem -v --stream-output --ignore-cache
> name                                           old time/op    new time/op    delta
> FlowSetup/vectorize=true/distribute=true-24       204µs ± 3%     210µs ± 4%  +2.89%  (p=0.000 n=28+27)
> FlowSetup/vectorize=true/distribute=false-24      205µs ± 8%     209µs ± 8%  +2.22%  (p=0.008 n=26+26)
> FlowSetup/vectorize=false/distribute=true-24      205µs ± 5%     210µs ± 6%  +2.57%  (p=0.000 n=29+30)
> FlowSetup/vectorize=false/distribute=false-24     201µs ± 6%     206µs ± 3%  +2.70%  (p=0.000 n=28+27)
>
> name                                           old alloc/op   new alloc/op   delta
> FlowSetup/vectorize=true/distribute=true-24      24.8kB ±13%    25.2kB ±13%  +1.49%  (p=0.005 n=29+29)
> FlowSetup/vectorize=true/distribute=false-24     23.6kB ±13%    23.4kB ±11%  -0.56%  (p=0.001 n=27+25)
> FlowSetup/vectorize=false/distribute=true-24     28.7kB ±10%    28.9kB ±11%    ~     (p=0.842 n=28+29)
> FlowSetup/vectorize=false/distribute=false-24    27.0kB ± 1%    27.0kB ± 1%    ~     (p=0.370 n=24+24)
>
> name                                           old allocs/op  new allocs/op  delta
> FlowSetup/vectorize=true/distribute=true-24         265 ±15%       273 ±14%  +2.81%  (p=0.001 n=30+30)
> FlowSetup/vectorize=true/distribute=false-24        253 ±17%       253 ± 1%  +0.10%  (p=0.001 n=27+23)
> FlowSetup/vectorize=false/distribute=true-24        262 ±10%       267 ±16%  +1.66%  (p=0.001 n=28+30)
> FlowSetup/vectorize=false/distribute=false-24       246 ± 1%       247 ± 0%  +0.51%  (p=0.000 n=24+19)
>
> (This is on 145bb3b
> <https://github.com/cockroachdb/cockroach/commit/145bb3b7bfa6508e280421d8ad2a841f41e869df>,
> and I manually modified scripts/bench to use 30 iterations instead of 10.)
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/cockroachdb/cockroach/pull/103866#issuecomment-1570946541>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ANA4FVAVQG2T35Q6X3VXLCDXI6WMDANCNFSM6AAAAAAYN3WY3A>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>