@knz 2023-06-19
> Interestingly, this seems to lead to an increase in allocations which actually increases the latency on the hot path:

Have we resolved this latency increase? I don't think this PR is bringing us forward if we keep this latency increase on the hot path.

Should I send a revert for this PR?