@miretskiy 2023-06-19
I reverted the change. Cancel checker is about 8x faster. But not 128x (or
1024x) faster

On Mon, Jun 19, 2023, 6:24 AM knz ***@***.***> wrote:

> Interestingly, this seems to lead to an increase in allocations which
> actually increases the latency on the hot path:
>
> Have we resolved this latency increase? I don't think this PR is bringing
> us forward if we keep this latency increase on the hot path.
>
> Should I send a revert for this PR?
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/cockroachdb/cockroach/pull/103866#issuecomment-1596925164>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ANA4FVGHC7RO4MU7JYLSFQDXMASE7ANCNFSM6AAAAAAYN3WY3A>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>